+++ src/children.cpp
@@ -8,25 +8,25 @@ Children::Children() {
   this->_parentId = 1;
 }
 
-Children::Children(int id, Date *date, std::string lastname, std::string firstname, float balance, int newParentId, e_type type, std::list<Record*> history) : Account(id, date, lastname, firstname, balance, type, history) {
+Children::Children(int id, Date *date, std::string lastname, std::string firstname, double balance, int newParentId, e_type type, std::list<Record*> history) : Account(id, date, lastname, firstname, balance, type, history) {
   this->_monthlyWithdraw = 0.0;
   this->_dailyWithdraw = 0.0;
   this->_parentId = newParentId;
 }
 
-float		Children::getMonthlyLimit() const {
+double		Children::getMonthlyLimit() const {
   return (this->_monthlyLimit);
 }
 
-float		Children::getDailyLimit() const {
+double		Children::getDailyLimit() const {
   return (this->_dailyLimit);
 }
 
-float		Children::getMonthlyWithdraw() const {
+double		Children::getMonthlyWithdraw() const {
   return (this->_monthlyWithdraw);
 }
 
-float		Children::getDailyWithdraw() const {
+double		Children::getDailyWithdraw() const {
   return (this->_dailyWithdraw);
 }
 
@@ -34,11 +34,11 @@ int		Children::getParentId() {
   return (this->_parentId);
 }
 
-void		Children::setMonthlyWithdraw(float	newMonthlyWithdraw) {
+void		Children::setMonthlyWithdraw(double	newMonthlyWithdraw) {
   this->_monthlyWithdraw = newMonthlyWithdraw;
 }
 
-void		Children::setDailyWithdraw(float	newDailyWithdraw) {
+void		Children::setDailyWithdraw(double	newDailyWithdraw) {
   this->_dailyWithdraw = newDailyWithdraw;
 }
 
@@ -46,9 +46,9 @@ void		Children::setParentId(int	newParentId) {
   this->_parentId = newParentId;
 }
 
-float		Children::withdraw(float amount) {
-  float	tmp_monthly = amount + this->_monthlyWithdraw;
-  float	tmp_daily = amount + this->_dailyWithdraw;
+double		Children::withdraw(double amount) {
+  double	tmp_monthly = amount + this->_monthlyWithdraw;
+  double	tmp_daily = amount + this->_dailyWithdraw;
   
   if (amount <= this->_balance && tmp_monthly <= this->_monthlyLimit && tmp_daily <= this->_dailyLimit) {
     if (this->_monthlyWithdraw <= this->_monthlyLimit && this->_dailyWithdraw <= this->_dailyLimit) {
