+++ src/record.cpp
@@ -6,7 +6,7 @@ Record::Record() {
   
 }
 
-Record::Record(Date *newDate, float newBalance) {
+Record::Record(Date *newDate, double newBalance) {
   this->_date = newDate;
   this->_balance = newBalance;
 }
@@ -15,7 +15,7 @@ Date	*Record::getDate() const {
   return (this->_date);
 }
 
-float	Record::getBalance() const {
+double	Record::getBalance() const {
   return (this->_balance);
 }
 
@@ -23,7 +23,7 @@ void	Record::setDate(Date *newDate) {
   this->_date = newDate;
 }
 
-void	Record::setBalance(float newBalance) {
+void	Record::setBalance(double newBalance) {
   this->_balance = newBalance;
 }
